feat: new service for presentation
- refactor: use the service in workflow
