+++ chk/infrastructure/view.py
@@ -5,6 +5,10 @@ import typing
 
 from pydantic import BaseModel
 
+from chk.infrastructure.file_loader import ExecuteContext
+from chk.infrastructure.helper import formatter
+from chk.infrastructure.symbol_table import ExecResponse
+
 
 class PresentationBuilder(BaseModel, abc.ABC):
     """Presentable signature"""
@@ -22,3 +26,20 @@ class PresentationBuilder(BaseModel, abc.ABC):
         """Signature to dump json"""
 
         raise NotImplementedError("Signature to dump json")
+
+
+class PresentationService:
+    """Comment service for presentation"""
+
+    @classmethod
+    def display(
+        cls,
+        ex_resp: ExecResponse,
+        exec_ctx: ExecuteContext,
+        presenter: type[PresentationBuilder],
+    ) -> None:
+        wfp = presenter(data=ex_resp)
+        if exec_ctx.options["format"]:
+            formatter(wfp.dump_fmt(), dump=exec_ctx.options["dump"])
+        else:
+            formatter(wfp.dump_json(), dump=exec_ctx.options["dump"])
