+++ chk/modules/workflow/__init__.py
@@ -16,7 +16,7 @@ from chk.infrastructure.document import (
     VersionedDocumentV2,
 )
 from chk.infrastructure.file_loader import ExecuteContext, FileContext
-from chk.infrastructure.helper import data_get, formatter, slugify
+from chk.infrastructure.helper import data_get, slugify
 from chk.infrastructure.symbol_table import (
     ExecResponse,
     ExposeManager,
@@ -25,7 +25,7 @@ from chk.infrastructure.symbol_table import (
     replace_value,
 )
 from chk.infrastructure.version import DocumentVersionMaker
-from chk.infrastructure.view import PresentationBuilder
+from chk.infrastructure.view import PresentationService
 from chk.modules.fetch import task_fetch
 from chk.modules.validate import task_validation
 from chk.modules.workflow.entities import (
@@ -174,19 +174,6 @@ class WorkflowDocumentSupport:
 
         return _task_res
 
-    @classmethod
-    def display(
-        cls,
-        ex_resp: ExecResponse,
-        exec_ctx: ExecuteContext,
-        presenter: type[PresentationBuilder],
-    ) -> None:
-        wfp = presenter(data=ex_resp)
-        if exec_ctx.options["format"]:
-            formatter(wfp.dump_fmt(), dump=exec_ctx.options["dump"])
-        else:
-            formatter(wfp.dump_json(), dump=exec_ctx.options["dump"])
-
 
 def call(file_ctx: FileContext, exec_ctx: ExecuteContext) -> ExecResponse:
     """Call a workflow document"""
@@ -232,4 +219,4 @@ def execute(
     exr = call(file_ctx=ctx, exec_ctx=exec_ctx)
 
     cb({ctx.filepath_hash: exr.variables_exec.data})
-    WorkflowDocumentSupport.display(exr, exec_ctx, WorkflowPresenter)
+    PresentationService.display(exr, exec_ctx, WorkflowPresenter)
